# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Care-Cordination server API documentation.",
        "title": "Care-Cordination API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/registrations": {
            "get": {
                "description": "List all registration forms",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "List registration forms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (pending, approved, rejected, in_review)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.PaginationResponse-registration_ListRegistrationFormsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new registration form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Create a registration form",
                "parameters": [
                    {
                        "description": "Registration Form",
                        "name": "registration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registration.CreateRegistrationFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registration.CreateRegistrationFormResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/registrations/{id}": {
            "get": {
                "description": "Get a registration form by ID with details",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Get a registration form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registration Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registration.GetRegistrationFormResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing registration form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Update a registration form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registration Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Registration Form Update",
                        "name": "registration",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/registration.UpdateRegistrationFormRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registration.UpdateRegistrationFormResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete a registration form by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Registration"
                ],
                "summary": "Delete a registration form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Registration Form ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/registration.DeleteRegistrationFormResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "registration.CreateRegistrationFormRequest": {
            "type": "object",
            "required": [
                "bsn",
                "careType",
                "dateOfBirth",
                "firstName",
                "gender",
                "lastName",
                "refferingOrgId",
                "registrationDate",
                "registrationReason"
            ],
            "properties": {
                "additionalNotes": {
                    "type": "string"
                },
                "attachmentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bsn": {
                    "type": "string"
                },
                "careType": {
                    "type": "string",
                    "enum": [
                        "protected_living",
                        "semi_independent_living",
                        "independent_assisted_living",
                        "ambulatory_care"
                    ]
                },
                "dateOfBirth": {
                    "type": "string",
                    "format": "2006-01-02"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "refferingOrgId": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string",
                    "format": "2006-01-02"
                },
                "registrationReason": {
                    "type": "string"
                }
            }
        },
        "registration.CreateRegistrationFormResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "registration.DeleteRegistrationFormResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "registration.GetRegistrationFormResponse": {
            "type": "object",
            "properties": {
                "additionalNotes": {
                    "type": "string"
                },
                "attachmentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bsn": {
                    "type": "string"
                },
                "careType": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hasClient": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "intakeCompleted": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "orgContactPerson": {
                    "type": "string"
                },
                "orgEmail": {
                    "type": "string"
                },
                "orgName": {
                    "type": "string"
                },
                "orgPhoneNumber": {
                    "type": "string"
                },
                "refferingOrgId": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "registrationReason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "registration.ListRegistrationFormsResponse": {
            "type": "object",
            "properties": {
                "additionalNotes": {
                    "type": "string"
                },
                "bsn": {
                    "type": "string"
                },
                "careType": {
                    "type": "string"
                },
                "dateOfBirth": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "intakeCompleted": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "numberOfAttachments": {
                    "type": "integer"
                },
                "orgContactPerson": {
                    "type": "string"
                },
                "orgEmail": {
                    "type": "string"
                },
                "orgName": {
                    "type": "string"
                },
                "orgPhoneNumber": {
                    "type": "string"
                },
                "refferingOrgId": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "registrationReason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "registration.UpdateRegistrationFormRequest": {
            "type": "object",
            "properties": {
                "additionalNotes": {
                    "type": "string"
                },
                "attachmentIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bsn": {
                    "type": "string"
                },
                "careType": {
                    "type": "string",
                    "enum": [
                        "protected_living",
                        "semi_independent_living",
                        "independent_assisted_living",
                        "ambulatory_care"
                    ]
                },
                "dateOfBirth": {
                    "type": "string",
                    "format": "2006-01-02"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female",
                        "other"
                    ]
                },
                "lastName": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "refferingOrgId": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string",
                    "format": "2006-01-02"
                },
                "registrationReason": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "pending",
                        "approved",
                        "rejected",
                        "in_review"
                    ]
                }
            }
        },
        "registration.UpdateRegistrationFormResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "resp.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resp.PaginationResponse-registration_ListRegistrationFormsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/registration.ListRegistrationFormsResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}